feat(school): implement school analysis ui

+++ lib/features/school/domain/school_analysis.dart
@@ -7,9 +7,9 @@ part 'school_analysis.g.dart';
 @freezed
 class SchoolAnalysis with _$SchoolAnalysis {
   const factory SchoolAnalysis({
-    @JsonKey(name: 'prevention_level') int? preventionLevel,
-    @JsonKey(name: 'emergency_response_level') int? emergencyResponseLevel,
-    @JsonKey(name: 'recovery_level') int? recoveryLevel,
+    @JsonKey(name: 'prevention_level') double? preventionLevel,
+    @JsonKey(name: 'emergency_response_level') double? emergencyResponseLevel,
+    @JsonKey(name: 'recovery_level') double? recoveryLevel,
   }) = _SchoolAnalysis;
 
   factory SchoolAnalysis.fromJson(Map<String, Object?> json) =>
